requirements.txt
Click==7.0
pwntools==3.12.2
PyYAML==3.12
